cheri: unintpr_t vs. unsigned long fixes in fs/ and include/net
More uintptr_t vs. unsigned long fixes in fs/ and include/net.

+++ fs/file.c
@@ -972,7 +972,7 @@ static inline struct file *__fget_files_rcu(struct files_struct *files,
 
 		/* Do the load, then mask any invalid result */
 		file = rcu_dereference_raw(*fdentry);
-		file = (void *)(nospec_mask & (unsigned long)file);
+		file = (void *)(nospec_mask & (uintptr_t)file);
 		if (unlikely(!file))
 			return NULL;
 
@@ -1128,7 +1128,7 @@ EXPORT_SYMBOL(task_lookup_next_fdget_rcu);
  * The fput_needed flag returned by fget_light should be passed to the
  * corresponding fput_light.
  */
-static unsigned long __fget_light(unsigned int fd, fmode_t mask)
+static uintptr_t __fget_light(unsigned int fd, fmode_t mask)
 {
 	struct files_struct *files = current->files;
 	struct file *file;
@@ -1146,21 +1146,21 @@ static unsigned long __fget_light(unsigned int fd, fmode_t mask)
 		file = files_lookup_fd_raw(files, fd);
 		if (!file || unlikely(file->f_mode & mask))
 			return 0;
-		return (unsigned long)file;
+		return (uintptr_t)file;
 	} else {
 		file = __fget_files(files, fd, mask);
 		if (!file)
 			return 0;
-		return FDPUT_FPUT | (unsigned long)file;
+		return FDPUT_FPUT | (uintptr_t)file;
 	}
 }
-unsigned long __fdget(unsigned int fd)
+uintptr_t __fdget(unsigned int fd)
 {
 	return __fget_light(fd, FMODE_PATH);
 }
 EXPORT_SYMBOL(__fdget);
 
-unsigned long __fdget_raw(unsigned int fd)
+uintptr_t __fdget_raw(unsigned int fd)
 {
 	return __fget_light(fd, 0);
 }
@@ -1181,9 +1181,9 @@ static inline bool file_needs_f_pos_lock(struct file *file)
 		(file_count(file) > 1 || file->f_op->iterate_shared);
 }
 
-unsigned long __fdget_pos(unsigned int fd)
+uintptr_t __fdget_pos(unsigned int fd)
 {
-	unsigned long v = __fdget(fd);
+	uintptr_t v = __fdget(fd);
 	struct file *file = (struct file *)(v & ~3);
 
 	if (file && file_needs_f_pos_lock(file)) {
